FIX Sql

+++ src/Domain/Herd/Entities/Herd.cs
@@ -2,10 +2,12 @@
 using HerdManagement.Domain.SpecieBreed.Entities;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
 namespace HerdManagement.Domain.Herd.Entities
 {
+    //[Table("Herd")]
     public class Herd : Entity<Herd>
     {
         public Herd()
@@ -16,7 +18,7 @@ namespace HerdManagement.Domain.Herd.Entities
         public string Name { get; set; }
         public string Description { get; set; }
         public string Color { get; set; }
-        public uint LivingMembersNumber { get; set; }
+        public long LivingMembersNumber { get; set; }
         public Specie Specie { get; set; }
 
         public int SpecieId { get; set; }
